product variants and similar products

+++ database/migrations/2025_04_06_095158_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('title');
             $table->text('content')->nullable();
-            $table->decimal('price')->default(0);
+            $table->integer('price')->default(0);
             $table->enum('discountType', ['percent', 'fixedSum'])->nullable();
             $table->integer('discountValue')->nullable();
             $table->integer('quantity')->default(0);
